+++ Sources/Superstring/AString.swift
@@ -13,18 +13,16 @@ import UIKit
 
 public struct AString: AttributedStringBuilder {
     
-    
-    public var components: (String, Attributes) {
+   public var components: AttributedStringComponents {
         (string, attributes)
     }
     
-    
-    public let string: String
-    public let attributes: Attributes
-    
     public var attributedString: NSAttributedString {
-        NSAttributedString(string: string, attributes: attributes)
+        NSAttributedString(string: components.string, attributes: components.attributes)
     }
+    
+    private let string: String
+    private let attributes: Attributes
     
     public init(_ string: String, attributes: Attributes = [:]) {
         self.string = string
@@ -33,7 +31,7 @@ public struct AString: AttributedStringBuilder {
     
     // MARK: - Internal
     
-   public func apply(_ newAttributes: Attributes) -> Self {
+   public func applying(_ newAttributes: Attributes) -> Self {
         var attributes = self.attributes
         // Merge the two dictionaries taking the value of the new attribute in case the key is duplicated.
         attributes.merge(newAttributes, uniquingKeysWith: { (_, new) in new } )