+++ Sources/Superstring/Newline.swift
@@ -10,14 +10,14 @@ import UIKit
 #endif
 
 public struct Newline: AttributedStringBuilder {
-    public var components: (String, Attributes) {
-        ("\n", attributes)
-    }
-    
     
     public var attributedString: NSAttributedString {
-        NSAttributedString(string: "\n", attributes: attributes)
+        NSAttributedString(string: components.string, attributes: components.attributes)
     }
+    
+    public var components: AttributedStringComponents {
+          ("\n", attributes)
+      }
     
     public let attributes: Attributes
     
@@ -27,7 +27,7 @@ public struct Newline: AttributedStringBuilder {
     
     // MARK: - Internal
     
-   public func apply(_ newAttributes: Attributes) -> Self {
+    public func applying(_ newAttributes: Attributes) -> Self {
         var attributes = self.attributes
         // Merge the two dictionaries taking the value of the new attribute in case the key is duplicated.
         attributes.merge(newAttributes, uniquingKeysWith: { (_, new) in new } )