+++ Sources/Superstring/Superstring.swift
@@ -11,44 +11,23 @@ import UIKit
 public struct Superstring: AttributedStringConvertible {
     
     public let attributedString: NSAttributedString
-    public let components: [String : Attributes] = [:]
-    public let componentss: [(String, Attributes)]
     
-    
-    public init(@SuperstringBuilder _ builder: () -> (AttributedStringConvertible, [(String, Attributes)])) {
-        attributedString = NSAttributedString(string: "sdaf")  //builder().attributedString
-        componentss = builder().1
-        
-        attributes  = [:]
-        print("🚨", attributedString)
-        print("🥎", componentss)
+    public init(_ attributes: Attributes = [:], @SuperstringBuilder _ builder: () -> [AttributedStringComponents]) {
+        attributedString = Self.makeAttributedString(from: builder(), attributes: attributes)
     }
     
-    let attributes: Attributes
+    public init(_ attributes: Attributes = [:], @SuperstringBuilder _ builder: () -> AttributedStringBuilder) {
+        attributedString = Self.makeAttributedString(from: [builder().components], attributes: attributes)
+    }
     
-    public init(_ attributes: Attributes, @SuperstringBuilder _ builder: () -> (AttributedStringConvertible, [(String, Attributes)])) {
-        self.attributes = attributes
-        
-        
-        componentss = builder().1
-        
-        let resultString = NSMutableAttributedString()
-     
-        
-        for c in componentss {
-            let atts = c.1.merging(attributes, uniquingKeysWith: { (c, new) in c } )
-            resultString.append(NSAttributedString(string: c.0, attributes: atts))
+    // MARK: - Private
+    
+    private static func makeAttributedString(from components: [AttributedStringComponents], attributes: Attributes) -> NSAttributedString {
+        let attributedString = NSMutableAttributedString()
+        components.forEach {
+            let resultingAttributes = $0.attributes.merging(attributes, uniquingKeysWith: { (current, _) in current } )
+            attributedString.append(NSAttributedString(string: $0.string, attributes: resultingAttributes))
         }
-        
-    
-        
-        
-          print("🥎", componentss)
-        attributedString = resultString  //builder().attributedString
+        return attributedString
     }
-    
-//    private static func makeAttributedString(_ string: String, baseAttributes: Attributes, attributes: Attributes) -> NSAttributedString {
-//        let at = baseAttributes.merge(attributes, uniquingKeysWith: { (_, new) in new } )
-//        let mutableString = NSMutableAttributedString()
-//    }
 }